pyproject.toml
[project]
name = "langgraph1"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chromadb>=1.5.5",
    "ipykernel>=7.2.0",
    "langchain>=1.2.10",
    "langchain-community>=0.4.1",
    "langchain-openai>=1.1.10",
    "langgraph>=1.0.10",
    "python-dotenv>=1.2.2",
]
